# public/api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dashboard Management API",
    "version": "1.0.0",
    "description": "A comprehensive API for user authentication and management with role-based permissions",
    "contact": {
      "name": "API Support",
      "email": "app@carik.id"
    }
  },
  "schemes": [
    "http",
    "https"
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login endpoint"
      }
    },
...
  },
  "paths": {
...
    "/v1/messages": {
      "post": {
        "summary": "Create a new message",
        "description": "Add a new message to an existing conversation",
        "tags": [
          "Messages"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Client-ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Client ID for the messages"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "conversation_id",
                  "content",
                  "role"
                ],
                "properties": {
                  "conversation_id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                  },
                  "parent_id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "123e4567-e89b-12d3-a456-426614174001"
                  },
                  "content": {
                    "type": "string",
                    "example": "This is a message content"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "user",
                      "assistant",
                      "system"
                    ],
                    "example": "user"
                  },
                  "model": {
                    "type": "string",
                    "example": "gpt-3.5-turbo"
                  },
                  "prompt_tokens": {
                    "type": "integer",
                    "example": 10
                  },
                  "completion_tokens": {
                    "type": "integer",
                    "example": 20
                  },
                  "total_tokens": {
                    "type": "integer",
                    "example": 30
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Message created successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "content": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string",
                          "enum": [
                            "user",
                            "assistant",
                            "system"
                          ]
                        },
                        "created_at": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/messages/streaming": {
      "post": {
        "summary": "Create a streaming assistant message",
        "description": "Create an empty assistant message for streaming updates",
        "tags": [
          "Messages"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Client-ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Client ID for the conversation"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "conversation_id"
                ],
                "properties": {
                  "conversation_id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                  },
                  "parent_id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "123e4567-e89b-12d3-a456-426614174001"
                  },
                  "model": {
                    "type": "string",
                    "example": "gpt-3.5-turbo"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Streaming message created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Streaming message created successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/messages/{id}": {
      "patch": {
        "summary": "Update a message",
        "description": "Update an existing message (used for streaming updates)",
        "tags": [
          "Messages"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Client-ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Client ID for the conversation"
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Message ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string",
                    "example": "Updated content"
                  },
                  "prompt_tokens": {
                    "type": "integer",
                    "example": 10
                  },
                  "completion_tokens": {
                    "type": "integer",
                    "example": 20
                  },
                  "total_tokens": {
                    "type": "integer",
                    "example": 30
                  },
                  "latency_ms": {
                    "type": "integer",
                    "example": 500
                  },
                  "status_id": {
                    "type": "integer",
                    "example": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Message updated successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}